% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fhs.R
\name{RollFHS}
\alias{RollFHS}
\title{Rolling Forecasts for Filtered Historical Simulation (FHS)}
\usage{
RollFHS(
  data,
  c,
  n,
  m,
  p = 1,
  q = 1,
  r = 1,
  model = "sGARCH",
  dist = "norm",
  type = c("EWMA", "GARCH"),
  lambda = 0.94,
  nboot = 10000,
  verbose = 0,
  solver = "hybrid",
  solver.control = list(tol = 1e-07),
  ...
)
}
\description{
Rolling Forecasts for Filtered Historical Simulation (FHS)
}
